rename fifocache to embedded cache
