+++ pkg/storage/chunk/cache/cache_test.go
@@ -205,7 +205,7 @@ func TestMemcache(t *testing.T) {
 }
 
 func TestFifoCache(t *testing.T) {
-	cache := cache.NewFifoCache("test", cache.FifoCacheConfig{MaxSizeItems: 1e3, TTL: 1 * time.Hour},
+	cache := cache.NewEmbeddedCache("test", cache.EmbeddedCacheConfig{MaxSizeItems: 1e3, TTL: 1 * time.Hour},
 		nil, log.NewNopLogger(), "test")
 	testCache(t, cache)
 }
